Moved this PR’s files up one package (out of `servlet`)
Considered renaming the whole repo to reflect the addition of these JPA stuff; but well, it kinda stuck. q:
