+++ xyz/paradoxv5/jpa/EntityManagerWrapper.java
@@ -1,4 +1,4 @@
-package xyz.paradoxv5.servlet.jpa;
+package xyz.paradoxv5.jpa;
 import java.util.function.Supplier;
 import javax.persistence.EntityManager;
 import javax.persistence.EntityTransaction;
