+++ xyz/paradoxv5/jpa/package-info.java
@@ -1,2 +1,2 @@
 /** Helper classes for accessing databases with the Java Persistent API (JPA) */
-package xyz.paradoxv5.servlet.jpa;
+package xyz.paradoxv5.jpa;
